requirements.txt
flask>=2.3.3
Flask-SQLAlchemy>=3.1.1